Cleaning up all shared properties and parameters
easy_type 0.7.4 allows you to define shared parameters and properties. This means that if properties are the same over some types, you only need to define them once and place them in the shared directory.

You must still include them with `parameter :shared_parameter` in the definition of the type.

Signed-off-by: Bert Hajee <hajee@moretIA.com>