Add debug capabilities to orawls
This commit adds the use of a temp directory to orawls. When you create
the following line in wls_settings on the host the module is ran,
orawls will copy all the scripts created as tempfiles to
/tmp/orawls-archive. This is useful when new or modified types are debugged.

debug_module: "true"

wls_settings needs to be changed, so we can puppetize this value as
well.