Add control over debug_module flag to wls_setting
As all configurable settings should be configured through Puppet, this
commit adds the debug_module flag to the wls_setting type. Default is
false. When switched to true, all the Jython scripts generated by the
module will be saved in /tmp/orawls-archive.

This functionality is fully backward compatible with earlier versions of
orawls and is deliberately not documented as it is most likely never
used by 'end users'