Add parameter archive_path and debug_mode
This commit adds the logic for the archive_path functionality. It also
adds the debug_mode ruby file, which I did not include in an earlier
commit.

All has been testes. Both parameters are optional, and you should not
need to provide them to use the module. They are very handy for
debugging though.