Add removeinfectedconnections datasource property
This new property will accept a boolean value as well as '0' and '1'.
It will manage the state of the 'Remove Infected Connections Enabled'
flag a datasource has.

This commit only implements the basic property fetching, munging etc.
I like getting this working properly first! :)

There is no default for the property which IMO is the correct behaviour.
Quoting *Puppet Types and Providers - Dan Bode and Nan Liu*
```
Defaults are typically only specified for parameters. They are less common for
properties because a default value will result in that property always being
managed (there is no way to unset a value for the property).
```