Added custom path to oracle inventory.
I made this because the installation of a OHS standalone server requires that the oracle inventory directory already exists. But if we keep this parameter inside oracle's installation path (for example /opt/oracle/oraInventory, with OHS being installed on /opt/oracle/middleware12c/...), when we define resources to create it, something in the puppet execution doesn't work as expected, due to some directory required by installation flow that it's already defined.