Running orawls with non-root user support
Add functionality to run orawls with a non-root user.
This will work
only with Puppet v3.7.0 or higher so as to use the patch for "exec"
resource type which does not switch user if running user & target user
are same & does not fail for switch user test. Not tested on
v4+
Following was done so as to have it running with non-root user:
1.
Override default values for wls_domains.yaml & wls_setting.yaml from
default /etc to custom location
2. For user parameters in "exec"
resource types, puppet has provided a patch from v3.7 onwards to
validate & check user if not root & does not fail if user to run as is
same as running user. So, for now, non-root user can only be run with
puppet v3.7+.
3. For bypassing easy_type running comman with (su -), the
code was modified to check it running user is root, else no user is
passed to easy_type daemon.rb