Modify the logic for when the memory params are changed
The current exec for changing the memory params will run unless the
ending condition is met. However, the command itself is dependant upon a
particular start condition. This leads to a loop where if the MEM_ARGS
don't match the 'unless', it will run the exec and the exec won't change
anything because sed isn't able to match anything (though it completes
successfully).

This PR will fix that by changing the 'unless' to an 'onlyif', and
making the condition of that 'onlyif' the same as the search string that
sed is looking for in the exec.